+++ apps/app/pages/users/[userId].js
@@ -1,5 +1,6 @@
-import Page from '@devlaunchers/user-profile/src/pages/users/[username]';
+import Page from '@devlaunchers/user-profile/src/pages/users/[userId].js';
 import App from '@devlaunchers/user-profile/src/pages/_app';
+export { getStaticPaths, getStaticProps} from '@devlaunchers/user-profile/src/pages/users/[userId].js'
 
 /////////////////////////////////////////
